+++ spec/requests/as_content_owner/video_spec.rb
@@ -427,7 +427,7 @@ describe Yt::Video, :partner do
         let(:keys) { Yt::Collections::Reports::SUBSCRIBED_STATUSES.keys }
 
         specify 'with the :by option set to subscribed statuses' do
-          views = video.views range.merge by: :subscribed_statuses
+          views = video.views range.merge by: :subscribed_status
           expect(views.keys - keys).to be_empty
           expect(views.values).to all(be_an Integer)
         end